Move tcpdump to internal package
This allows us to reuse it for the upcoming bpfoff command, without it
being publicly exported and having to worry about APIs and what not.
